env: sf: Try to load the saved environment after CRC errors by repeatedly reading
This is a hack that addresses problems of U-Boot loading the environment during the boot process
in DRC02 baseboards. A Cypress MCU is used there and it boots parallel to the DHCOM i.MX6 and
(seems to) block the SPI bus for a short moment (sometimes).
To this SPI bus the bootflash and the Cypress MCU are connected as slave (2 CS lines).