dh_imx6: boot from SD card if barebox found
WashTec's Multibox needs to be migrated from U-Boot/WEC7 to
barebox/linux in the field. In order to keep things simple and flexible
on U-Boot's side, simply check for a barebox header on the SD card (if
inserted). If such a header is found, switch to SD card boot and reset.

The rest of the migration happens in the SD card image barebox is part
of (including overwriting U-Boot in the eMMC boot partition and WEC7 on
the eMMC user partition).

Signed-off-by: Bastian Krause <bst@pengutronix.de>